refactor(chatbot): platform-neutral gateway poller and message handler
The inbound gateway poller and its message handler were YouTube-named
but already platform-generic — the gateway owns every platform-specific
concern (discovery, paging, own-echo filtering, cadence). Rename them
gatewayChatPoller / HandleGatewayMessage and move them to
gateway_chat.go so a second gateway-wired platform reuses them; the
constructor takes the gateway base URL instead of reading the YouTube
one. youtube.go keeps the YouTube-specific outbound client.
